Expand label on migrated_ref_time
Variable migrated_ref_time was "Reference time applied to migration questions". It implicitly was thought as years, as all questionnaires seen or thought about were in years, but in some cases it actually is months. So this should be made clear in the label.

For example, a 6 months period would thus be 0.5.