#Section 4.1 Static Checks GLD
Hi Mario, I think that introducing this line will fix a recurrent issue from the checks. In the output I used to get for all my datasets issues with the hhsize and number of obs per household. After a quick br, I noticed that stata generated integers(7, 4, 5, 3) for the variable hh_size_verify even though hsize had missing values for the hhid. I fixed this so that the count now compares equivalent variables. I am happy to discuss this further if you wish.