ISCO/SINCO R code
Hi Mario, 
I changed some of the lines of the original ISIC code to create the ISCO/SINCO code using the correspondence table from the INEGI.
Basically, I commented out sections 5 and 6, and ran the code replacing some of the lines to make sure the data frame specifications fit the data of the excel file. 
I would be happy to hear your thoughts on this version to see if we can use it as the final for SINC/ ISCO version for matching. The CMO is other story that may require a separate discussion.