Update external urban check
Code now drops cases of urban that are missing. Idea is that, the degree of "missingness" is reported elsewhere, here we want to know whether the remaining info is in line with international estimates. If missing at random, should have no influence. If missing are all urban, then this should reflect here in being lopsided.